Split files to bonus
