+++ maps/3.ber
@@ -1,0 +1,14 @@
+11111111111111111111111111111111111111111111111111111111
+1P0C1111111111111111111111111111111111111111111111111111
+11000000000000000000000000000000000000000000000000000011
+11000000000000000000000000000000000000000000000000000011
+11000000000000000000000000000000000000000000000000000011
+11000000000000000000000000000000000000000000000000000011
+10000000000000000000000000000000000000000000000000000001
+10000000000000000000000000000000000000000000000000000001
+10000000000000000000000000000000000000000000000000000001
+10000000000000000000000000000000000000000000000000000001
+10000000000000000000000000000000000000000000000000000001
+10000000000000000000000000000000000000000000000000000001
+100000000000000000000000000000000000000000000000000000E1
+11111111111111111111111111111111111111111111111111111111
