+++ maps/invalid_3.ber
@@ -1,0 +1,1 @@
+
