+++ src/DFS.c
@@ -6,7 +6,7 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 19:39:36 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/07 10:12:13 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 17:36:03 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -104,18 +104,3 @@ bool	check_path(t_data *data, char **map)
 			data->game->player->y);
 	return (ft_free_2d_arr(visited), res);
 }
-
-// int main(int argc, char *argv[])
-// {
-// 	t_data	data;
-
-// 	if (argc != 2)
-// 		return (1);
-// 	data.game.map.fd = open(argv[1], O_RDONLY);
-// 	read_map(&data);
-//     if (check_path(data.game.map.full))
-//         printf("You have reached the exit\nThere is indeed a valid path.\n");
-//     else
-//         printf("All paths checked. No path to exit found\n");
-//     return (0);
-// }
