+++ src/free.c
@@ -6,7 +6,7 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 14:52:29 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/14 15:39:15 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 17:36:27 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,6 @@ void free_textures(t_data *data)
 {
 	if (have_textures_loaded(data))
 	{
-		free_sprite(data, data->game->textures->enemy);
-		free_sprite(data, data->game->textures->wall);
 		free_sprite(data, data->game->textures->border_0);
 		free_sprite(data, data->game->textures->border_1);
 		free_sprite(data, data->game->textures->border_2);
@@ -31,6 +29,7 @@ void free_textures(t_data *data)
 		free_sprite(data, data->game->textures->border_5);
 		free_sprite(data, data->game->textures->border_6);
 		free_sprite(data, data->game->textures->border_7);
+		free_sprite(data, data->game->textures->wall);
 		free_sprite(data, data->game->textures->collectible);
 		free_sprite(data, data->game->textures->empty);
 		free_sprite(data, data->game->textures->exit);
