+++ src/init_textures.c
@@ -6,11 +6,23 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/08 15:53:50 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/14 14:08:18 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 17:03:52 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/so_long.h"
+
+void	*create_texture(t_data data, char *file_name)
+{
+	void	*texture;
+	int		height;
+	int		width;
+
+	height = PLAYER_HEIGHT;
+	width = PLAYER_WIDTH;
+	texture = mlx_xpm_file_to_image(data.mlx_ptr, file_name, &width, &height);
+	return (texture);
+}
 
 bool have_textures_loaded(t_data *data)
 {
