+++ src/movement.c
@@ -6,7 +6,7 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/05 14:34:03 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/09 12:11:38 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 17:03:02 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,14 +45,6 @@ int	move_to(t_data *data, int new_x, int new_y)
 		++data->game->collectables->collected;
 	else if (data->game->map->full[new_y][new_x] == 'E')
 		return (can_exit(data));
-	else if (data->game->enemy
-		&& new_y == data->game->enemy->path->current_tile->y
-		&& new_x == data->game->enemy->path->current_tile->x)
-	{
-		ft_putstr_endl("You died bruh", 1);
-		handle_destroy(data);
-		return (0);
-	}
 	move_player(data, new_x, new_y);
 	return (0);
 }
