+++ src_bonus/DFS.c
@@ -6,11 +6,11 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 19:39:36 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/07 10:12:13 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 21:00:42 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/so_long.h"
+#include "../includes/so_long_bonus.h"
 
 /**
  * @brief Get the size of a 2d arr. Relies on NULL Terminator
