+++ src_bonus/args.c
@@ -6,11 +6,11 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 14:50:17 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/05 15:54:08 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 21:00:37 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/so_long.h"
+#include "../includes/so_long_bonus.h"
 
 static bool	is_ext_valid(char *file)
 {
