+++ src_bonus/enemy.c
@@ -6,11 +6,11 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 14:54:01 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/09 12:11:47 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 21:00:48 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/so_long.h"
+#include "../includes/so_long_bonus.h"
 
 void	move_enemy(t_data *data)
 {
