+++ src_bonus/init_sprite_animated.c
@@ -6,11 +6,11 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/18 16:28:02 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/09 12:04:47 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 21:00:56 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/so_long.h"
+#include "../includes/so_long_bonus.h"
 
 char	*create_file_name(char *dir, int i)
 {
