+++ src_bonus/render_utils.c
@@ -6,11 +6,11 @@
 /*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/07 15:53:20 by maabdull          #+#    #+#             */
-/*   Updated: 2024/03/09 12:04:19 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/03/14 21:01:22 by maabdull         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/so_long.h"
+#include "../includes/so_long_bonus.h"
 
 void	render_texture(t_data *data, void *texture, int x, int y)
 {
